ios: share reducers and mark bridge plan progress
